Associate flavor types with datastore versions
This feature provides the ability to associate flavor types with datastore versions.
The management api will provide the ability to add or delete this information.
The normal user can view the flavor list for the default version or a
specified datastore version id. The troveclient has been changed to provide the cli
support for this (list flavors).

Change-Id: I89fc2db0078d7884b7b3fa91f5daec68c0132dfb
Implements: blueprint associate-flavors-datastores